TvTennisCDM8 0.01
- Representing Ball Velocity (page 6)

not sure the reasoning for this but i have followed the instructions
from page 6 on the project proposal

Representing ball velocity:
The velocity, vxy, of the ball has a horizontal (x) component, vx, and
a vertical (y)
component, vy. Each of these is represented by a 3-bit 2’s complement
(signed) number,
meaning that vx and vy can each take on values in the range -4..+3.
This is a very coarse
scale, but it is good enough for our purposes.
A positive x velocity means that the ball is travelling from left to
right across the space, and
a positive y velocity means that the ball is travelling from bottom to
top. Negative velocities
of course mean that the ball is travelling in the opposite direction.
Any combination of vx
and vy is possible.
Every data item read by the CdM-8 core must be represented as an 8-bit
string, because
CdM-8 is an 8-bit platform. Rather than pad each of the two velocity
components to 8 bits
we combine them into one 8-bit string as follows: bits 0..2 carry vx,
bits 3..5 carry vy, and
bits 6..7 are always set to 0.